Move some network conformance tests around.
Move the Endpoints API test from endpointslice.go to endpoints.go

Move the "kubernetes.default Service exists" and "kubernetes.default
endpoints exist" tests to apiserver.go, since (unlike the rest of
service.go/endpointslice.go) they aren't testing the behavior of the
Service/EndpointSlice/Endpoints APIs.

(No code changes, but fixed a typo in a comment.)
