+++ test/e2e/network/endpointslice.go
@@ -61,39 +61,6 @@ var _ = common.SIGDescribe("EndpointSlice", func() {
 
 	/*
 		Release: v1.21
-		Testname: kubernetes.default Endpoints and EndpointSlices
-		Description: The "kubernetes.default" service MUST have Endpoints and EndpointSlices pointing to each API server instance.
-	*/
-	framework.ConformanceIt("should have Endpoints and EndpointSlices pointing to API Server", func(ctx context.Context) {
-		namespace := "default"
-		name := "kubernetes"
-		// verify "kubernetes.default" service exist
-		_, err := cs.CoreV1().Services(namespace).Get(ctx, name, metav1.GetOptions{})
-		framework.ExpectNoError(err, "error obtaining API server \"kubernetes\" Service resource on \"default\" namespace")
-
-		// verify Endpoints for the API servers exist
-		endpoints, err := cs.CoreV1().Endpoints(namespace).Get(ctx, name, metav1.GetOptions{})
-		framework.ExpectNoError(err, "error obtaining API server \"kubernetes\" Endpoint resource on \"default\" namespace")
-		if len(endpoints.Subsets) == 0 {
-			framework.Failf("Expected at least 1 subset in endpoints, got %d: %#v", len(endpoints.Subsets), endpoints.Subsets)
-		}
-		// verify EndpointSlices for the API servers exist
-		endpointSliceList, err := cs.DiscoveryV1().EndpointSlices(namespace).List(ctx, metav1.ListOptions{
-			LabelSelector: "kubernetes.io/service-name=" + name,
-		})
-		framework.ExpectNoError(err, "error obtaining API server \"kubernetes\" EndpointSlice resource on \"default\" namespace")
-		if len(endpointSliceList.Items) == 0 {
-			framework.Failf("Expected at least 1 EndpointSlice, got %d: %#v", len(endpoints.Subsets), endpoints.Subsets)
-		}
-
-		if !endpointSlicesEqual(endpoints, endpointSliceList) {
-			framework.Failf("Expected EndpointSlice to have same addresses and port as Endpoints, got %#v: %#v", endpoints, endpointSliceList)
-		}
-
-	})
-
-	/*
-		Release: v1.21
 		Testname: EndpointSlice, "empty" Service
 		Description: The EndpointSlice controller should create and delete empty EndpointSlices for a Service that matches no pods.
 	*/
@@ -1016,58 +983,4 @@ func ensurePodTargetRef(pod *v1.Pod, targetRef *v1.ObjectReference) {
 	if targetRef.UID != pod.UID {
 		framework.Failf("Expected TargetRef.UID to be %s, got %s", pod.UID, targetRef.UID)
 	}
-}
-
-// endpointSlicesEqual compare if the Endpoint and the EndpointSliceList contains the same endpoints values
-// as in addresses and ports, considering Ready and Unready addresses
-func endpointSlicesEqual(endpoints *v1.Endpoints, endpointSliceList *discoveryv1.EndpointSliceList) bool {
-	// get the apiserver endpoint addresses
-	epAddresses := sets.NewString()
-	epPorts := sets.NewInt32()
-	for _, subset := range endpoints.Subsets {
-		for _, addr := range subset.Addresses {
-			epAddresses.Insert(addr.IP)
-		}
-		for _, addr := range subset.NotReadyAddresses {
-			epAddresses.Insert(addr.IP)
-		}
-		for _, port := range subset.Ports {
-			epPorts.Insert(port.Port)
-		}
-	}
-	framework.Logf("Endpoints addresses: %v , ports: %v", epAddresses.List(), epPorts.List())
-
-	// Endpoints are single stack, and must match the primary IP family of the Service kubernetes.default
-	// However, EndpointSlices can be IPv4 or IPv6, we can only compare the Slices that match the same IP family
-	// framework.TestContext.ClusterIsIPv6() reports the IP family of the kubernetes.default service
-	var addrType discoveryv1.AddressType
-	if framework.TestContext.ClusterIsIPv6() {
-		addrType = discoveryv1.AddressTypeIPv6
-	} else {
-		addrType = discoveryv1.AddressTypeIPv4
-	}
-
-	// get the apiserver addresses from the endpoint slice list
-	sliceAddresses := sets.NewString()
-	slicePorts := sets.NewInt32()
-	for _, slice := range endpointSliceList.Items {
-		if slice.AddressType != addrType {
-			framework.Logf("Skipping slice %s: wanted %s family, got %s", slice.Name, addrType, slice.AddressType)
-			continue
-		}
-		for _, s := range slice.Endpoints {
-			sliceAddresses.Insert(s.Addresses...)
-		}
-		for _, ports := range slice.Ports {
-			if ports.Port != nil {
-				slicePorts.Insert(*ports.Port)
-			}
-		}
-	}
-
-	framework.Logf("EndpointSlices addresses: %v , ports: %v", sliceAddresses.List(), slicePorts.List())
-	if sliceAddresses.Equal(epAddresses) && slicePorts.Equal(epPorts) {
-		return true
-	}
-	return false
 }
